# docs-site/contracts/product-http.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "WuKongIM HTTP API",
    "version": "3.0.0-beta-snapshot",
    "description": "The 48 core WuKongIM HTTP API JSON operations, plus the plugin-defined ANY route in x-wukongim-extension-routes. Routes have no built-in authentication.",
    "x-i18n": {
      "zh": {
        "title": "WuKongIM HTTP API",
        "description": "与当前运行时的 48 个核心 WuKongIM HTTP API JSON 操作对齐；插件定义的 ANY 路由见 x-wukongim-extension-routes。这些路由没有内建鉴权。"
      }
    }
  },
  "servers": [
    {
      "url": "http://127.0.0.1:5001",
      "description": "Local single-node cluster example",
      "x-i18n": {
        "zh": {
          "description": "本地单节点集群示例"
        }
      }
    }
  ],
...
  "security": [],
  "paths": {
...
    "/message/send": {
      "post": {
        "operationId": "sendChannelMessage",
        "summary": "Submit a message",
        "description": "Accepts the complete compatibility parser, including legacy aliases, transient flags, and request-scoped subscribers; inspect reason on HTTP 200.",
        "x-i18n": {
          "zh": {
            "summary": "提交消息",
            "description": "接受完整兼容 parser，包括旧别名、瞬时标志与请求级订阅者；HTTP 200 时仍需检查 reason。"
          }
        },
        "tags": [
          "Message Sending"
        ],
        "security": [],
        "x-wukongim-trust": "trusted-backend-only",
        "parameters": [
          {
            "$ref": "#/components/parameters/TraceID"
          }
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/SendMessage"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/SendCompleted"
          },
          "400": {
            "$ref": "#/components/responses/SendBadRequest"
          },
          "404": {
            "$ref": "#/components/responses/SendNotFound"
          },
          "408": {
            "$ref": "#/components/responses/SendTimedOut"
          },
          "500": {
            "$ref": "#/components/responses/SendInternalError"
          },
          "503": {
            "$ref": "#/components/responses/SendUnavailable"
          }
        }
      }
    },
    "/message/event": {
      "post": {
        "operationId": "appendMessageEvent",
        "summary": "Append a message event",
        "description": "Validates and applies one message-scoped event projection; non-null headers are unsupported.",
        "x-i18n": {
          "zh": {
            "summary": "追加消息事件",
            "description": "校验并应用一次消息级事件投影；不支持非 null headers。"
          }
        },
        "tags": [
          "Messages"
        ],
        "security": [],
        "x-wukongim-trust": "trusted-backend-only",
        "requestBody": {
          "$ref": "#/components/requestBodies/AppendMessageEvent"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/MessageEventOK"
          },
          "400": {
            "$ref": "#/components/responses/CompatibilityBadRequest"
          },
          "503": {
            "$ref": "#/components/responses/MaintenanceUnavailable"
          }
        }
      }
    },
    "/message/eventsync": {
      "post": {
        "operationId": "syncMessageEvents",
        "summary": "Synchronize current message event projections",
        "description": "Reads current durable event projections after a sequence cursor, with filtering after a bounded read; not a replayable event log.",
        "x-i18n": {
          "zh": {
            "summary": "同步消息事件当前投影",
            "description": "按序号游标读取当前持久化事件投影，不重放完整事件日志。在有界读取后过滤私有事件。调用方身份与 include_private 没有鉴权，必须由受信后端保护。"
          }
        },
        "tags": [
          "Messages"
        ],
        "security": [],
        "x-wukongim-trust": "trusted-backend-only",
        "requestBody": {
          "$ref": "#/components/requestBodies/MessageEventSync"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/MessageEventSyncOK"
          },
          "400": {
            "$ref": "#/components/responses/CompatibilityBadRequest"
          },
          "503": {
            "$ref": "#/components/responses/MaintenanceUnavailable"
          }
        }
      }
    },
...
    "/message/cmd/bind": {
      "post": {
        "operationId": "bindCommandChannel",
        "summary": "Bind command-channel discovery",
        "description": "Starts bounded recipient discovery after one committed command tail.",
        "x-i18n": {
          "zh": {
            "summary": "绑定命令 Channel 发现",
            "description": "为有上限的接收者批次建立 CMD 发现关系，从同一已提交尾部之后生效。"
          }
        },
        "tags": [
          "Messages"
        ],
        "deprecated": true,
        "security": [],
        "x-wukongim-trust": "trusted-backend-only",
        "requestBody": {
          "$ref": "#/components/requestBodies/MessageCMDBinding"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/StatusOK"
          },
          "400": {
            "$ref": "#/components/responses/CompatibilityBadRequest"
          },
          "503": {
            "$ref": "#/components/responses/MaintenanceUnavailable"
          }
        }
      }
    },
    "/message/cmd/unbind": {
      "post": {
        "operationId": "unbindCommandChannel",
        "summary": "Unbind command-channel discovery",
        "description": "Tombstones bounded recipient discovery without deleting command messages.",
        "x-i18n": {
          "zh": {
            "summary": "解绑命令 Channel 发现",
            "description": "为有上限的接收者批次写入发现关系墓碑，不删除 CMD 消息。"
          }
        },
        "tags": [
          "Messages"
        ],
        "deprecated": true,
        "security": [],
        "x-wukongim-trust": "trusted-backend-only",
        "requestBody": {
          "$ref": "#/components/requestBodies/MessageCMDBinding"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/StatusOK"
          },
          "400": {
            "$ref": "#/components/responses/CompatibilityBadRequest"
          },
          "503": {
            "$ref": "#/components/responses/MaintenanceUnavailable"
          }
        }
      }
    },
    "/channel/messagesync": {
      "post": {
        "operationId": "syncQuickstartChannelMessages",
        "summary": "Synchronize one Channel's committed messages",
        "description": "Checks membership visibility and returns an ascending page; limit is 100 by default and capped at 10000.",
        "x-i18n": {
          "zh": {
            "summary": "同步一个 Channel 的已提交消息",
            "description": "校验成员可见性并返回升序分页；limit 默认 100，最大 10000。"
          }
        },
        "tags": [
          "Messages"
        ],
        "security": [],
        "x-wukongim-trust": "trusted-backend-only",
        "requestBody": {
          "$ref": "#/components/requestBodies/ChannelMessageSync"
        },
        "responses": {
          "200": {
            "description": "One committed-message page",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChannelMessageSyncResponse"
                }
              }
            },
            "headers": {
              "X-WK-Content-Epoch": {
                "schema": {
                  "type": "string",
                  "pattern": "^[0-9]+$"
                },
                "description": "Successful-restore generation. Invalidate cached comparisons when it increases; discard delayed older generations.",
                "x-i18n": {
                  "zh": {
                    "description": "成功恢复代数；增加后重置缓存版本比较，丢弃延迟到达的旧代数响应。"
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/CompatibilityBadRequest"
          },
          "503": {
            "description": "Temporary read failure or restore maintenance; retry the same query with bounded backoff.",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/MaintenanceError"
                    },
                    {
                      "$ref": "#/components/schemas/MessageUpdateError"
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/channel/messagesyncbatch": {
      "post": {
        "operationId": "syncChannelMessagesBatch",
        "summary": "Synchronize up to 200 Channels",
        "description": "Validates all memberships before one aligned batch read; item failures are embedded in the HTTP-200 response.",
        "x-i18n": {
          "zh": {
            "summary": "同步最多 200 个 Channel",
            "description": "批量读取前校验全部成员关系；单项失败嵌入 HTTP 200 响应。"
          }
        },
        "tags": [
          "Messages"
        ],
        "security": [],
        "x-wukongim-trust": "trusted-backend-only",
        "requestBody": {
          "$ref": "#/components/requestBodies/ChannelMessageSyncBatch"
        },
        "responses": {
          "200": {
            "description": "Aligned batch result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChannelMessageSyncBatchResponse"
                }
              }
            },
            "headers": {
              "X-WK-Content-Epoch": {
                "schema": {
                  "type": "string",
                  "pattern": "^[0-9]+$"
                },
                "description": "Successful-restore generation. Invalidate cached comparisons when it increases; discard delayed older generations.",
                "x-i18n": {
                  "zh": {
                    "description": "成功恢复代数；增加后重置缓存版本比较，丢弃延迟到达的旧代数响应。"
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/CompatibilityBadRequest"
          },
          "503": {
            "description": "Content epoch unavailable, restore transition, or restore maintenance; existing per-item batch errors are unchanged.",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/MaintenanceError"
                    },
                    {
                      "$ref": "#/components/schemas/MessageUpdateError"
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
...
    "/messages": {
      "post": {
        "operationId": "lookupMessages",
        "summary": "Look up exact committed messages",
        "description": "Looks up exact committed messages within membership visibility; selector, result and scan limits fail without partial results.",
        "x-i18n": {
          "zh": {
            "summary": "精确查询已提交消息",
            "description": "按成员可见范围精确查询已提交消息；选择项、结果或索引检查超限时明确失败，不返回部分结果。"
          }
        },
        "tags": [
          "Messages"
        ],
        "security": [],
        "x-wukongim-trust": "trusted-backend-only",
        "requestBody": {
          "$ref": "#/components/requestBodies/MessageLookup"
        },
        "responses": {
          "200": {
            "description": "One committed-message page",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChannelMessageSyncResponse"
                }
              }
            },
            "headers": {
              "X-WK-Content-Epoch": {
                "schema": {
                  "type": "string",
                  "pattern": "^[0-9]+$"
                },
                "description": "Successful-restore generation. Invalidate cached comparisons when it increases; discard delayed older generations.",
                "x-i18n": {
                  "zh": {
                    "description": "成功恢复代数；增加后重置缓存版本比较，丢弃延迟到达的旧代数响应。"
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/CompatibilityBadRequest"
          },
          "503": {
            "description": "Temporary read failure or restore maintenance; retry the same query with bounded backoff.",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/MaintenanceError"
                    },
                    {
                      "$ref": "#/components/schemas/MessageUpdateError"
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/message/update": {
      "post": {
        "operationId": "updateMessage",
        "summary": "Replace a message payload",
        "description": "Replaces an ordinary retained payload with version and restore-epoch checks. CMD, stream and nonpersistent messages cannot be edited.",
        "x-i18n": {
          "zh": {
            "summary": "修改消息内容",
            "description": "按版本和恢复代数校验替换普通保留消息；CMD、流消息和非持久化消息不可修改。"
          }
        },
        "tags": [
          "Messages"
        ],
        "security": [],
        "x-wukongim-trust": "trusted-backend-only",
        "requestBody": {
          "$ref": "#/components/requestBodies/UpdateMessage"
        },
        "responses": {
          "200": {
            "description": "Success",
            "headers": {
              "X-WK-Content-Epoch": {
                "schema": {
                  "type": "string",
                  "pattern": "^[0-9]+$"
                },
                "description": "Successful-restore generation. Invalidate cached comparisons when it increases; discard delayed older generations.",
                "x-i18n": {
                  "zh": {
                    "description": "成功恢复代数；增加后重置缓存版本比较，丢弃延迟到达的旧代数响应。"
                  }
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateMessageResponse"
                }
              }
            }
          },
          "400": {
            "description": "invalid_request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageUpdateError"
                }
              }
            }
          },
          "403": {
            "description": "channel_not_accessible",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageUpdateError"
                }
              }
            }
          },
          "404": {
            "description": "message_not_found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageUpdateError"
                }
              }
            }
          },
          "409": {
            "description": "version_conflict, content_epoch_conflict, idempotency_conflict, reset_required or stale_meta",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageUpdateError"
                }
              }
            }
          },
          "422": {
            "description": "message_not_updatable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageUpdateError"
                }
              }
            }
          },
          "429": {
            "description": "resource_exhausted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageUpdateError"
                }
              }
            }
          },
          "503": {
            "description": "Temporary dependency failure or restore maintenance",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/MaintenanceError"
                    },
                    {
                      "$ref": "#/components/schemas/MessageUpdateError"
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/channel/messageupdates": {
      "post": {
        "operationId": "syncMessageUpdates",
        "summary": "Pull edits for one channel",
        "description": "Bootstrap with an empty cursor before loading history. Drain more pages and merge versions with the cursor atomically; this feed returns latest states.",
        "x-i18n": {
          "zh": {
            "summary": "拉取单频道消息修改",
            "description": "先用空游标获取基线，再加载历史；拉完 more 分页，在同一事务中合并版本并保存游标。此接口返回最新状态。"
          }
        },
        "tags": [
          "Messages"
        ],
        "security": [],
        "x-wukongim-trust": "trusted-backend-only",
        "requestBody": {
          "$ref": "#/components/requestBodies/MessageUpdates"
        },
        "responses": {
          "200": {
            "description": "Success",
            "headers": {
              "X-WK-Content-Epoch": {
                "schema": {
                  "type": "string",
                  "pattern": "^[0-9]+$"
                },
                "description": "Successful-restore generation. Invalidate cached comparisons when it increases; discard delayed older generations.",
                "x-i18n": {
                  "zh": {
                    "description": "成功恢复代数；增加后重置缓存版本比较，丢弃延迟到达的旧代数响应。"
                  }
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageUpdatesResponse"
                }
              }
            }
          },
          "400": {
            "description": "invalid_request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageUpdateError"
                }
              }
            }
          },
          "403": {
            "description": "channel_not_accessible",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageUpdateError"
                }
              }
            }
          },
          "404": {
            "description": "message_not_found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageUpdateError"
                }
              }
            }
          },
          "409": {
            "description": "version_conflict, content_epoch_conflict, idempotency_conflict, reset_required or stale_meta",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageUpdateError"
                }
              }
            }
          },
          "422": {
            "description": "message_not_updatable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageUpdateError"
                }
              }
            }
          },
          "429": {
            "description": "resource_exhausted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageUpdateError"
                }
              }
            }
          },
          "503": {
            "description": "Temporary dependency failure or restore maintenance",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/MaintenanceError"
                    },
                    {
                      "$ref": "#/components/schemas/MessageUpdateError"
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "parameters": {
...
      "TraceID": {
        "name": "X-WK-Trace-ID",
        "in": "header",
        "required": false,
        "description": "Optional 32-character hexadecimal trace ID. Invalid values are ignored and replaced with a generated ID.",
        "schema": {
          "type": "string",
          "pattern": "^[0-9a-fA-F]{32}$"
        }
      }
    },
    "requestBodies": {
...
      "SendMessage": {
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/SendMessageRequest"
            }
          }
        }
      },
      "MessageEventSync": {
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/MessageEventSyncRequest"
            }
          }
        }
      },
      "AppendMessageEvent": {
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/AppendMessageEventRequest"
            }
          }
        }
      },
...
      "MessageCMDBinding": {
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/MessageCMDBindingRequest"
            }
          }
        }
      },
      "ChannelMessageSync": {
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ChannelMessageSyncRequest"
            }
          }
        }
      },
...
    },
    "responses": {
      "StatusOK": {
        "description": "Mutation accepted",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/StatusEnvelope"
            }
          }
        }
      },
      "CompatibilityBadRequest": {
        "description": "Input, validation, dependency, or downstream error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/CompatibilityError"
            }
          }
        }
      },
      "MaintenanceUnavailable": {
        "description": "Controller restore maintenance is active",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/MaintenanceError"
            }
          }
        }
      },
...
      "SendCompleted": {
        "description": "Transport completed; inspect reason",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/SendMessageResponse"
            }
          }
        }
      },
      "SendBadRequest": {
        "description": "Invalid request or payload",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/SendError"
            }
          }
        }
      },
      "SendNotFound": {
        "description": "Channel not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/SendError"
            }
          }
        }
      },
      "SendTimedOut": {
        "description": "Request cancelled or timed out",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/SendError"
            }
          }
        }
      },
      "SendInternalError": {
        "description": "Unmapped send failure",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/SendError"
            }
          }
        }
      },
      "SendUnavailable": {
        "description": "Retry-required routing failure or restore maintenance",
        "content": {
          "application/json": {
            "schema": {
              "oneOf": [
                {
                  "$ref": "#/components/schemas/RetryRequiredError"
                },
                {
                  "$ref": "#/components/schemas/MaintenanceError"
                }
              ]
            }
          }
        }
      },
      "MessageEventSyncOK": {
        "description": "Current durable event projections",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/MessageEventSyncResponse"
            }
          }
        }
      },
      "MessageEventOK": {
        "description": "Applied event projection",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/AppendMessageEventResponse"
            }
          }
        }
      },
...
    },
    "schemas": {
...
      "ChannelMessageSyncResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "start_message_seq",
          "end_message_seq",
          "more",
          "messages"
        ],
        "properties": {
          "start_message_seq": {
            "type": "integer",
            "format": "uint64",
            "minimum": 0
          },
          "end_message_seq": {
            "type": "integer",
            "format": "uint64",
            "minimum": 0
          },
          "more": {
            "type": "integer",
            "enum": [
              0,
              1
            ]
          },
          "messages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LegacyMessage"
            }
          }
        }
      },
...
    }
  }
}